addons/pinniped/config-controller: remove need for deletecollection RBAC
We were seeing this error on our DeleteAllOf() calls. We can get the name of the
Secret for the Cluster, so just Delete() that single Secret. I also feel like in
doing it this way, we have less of a chance of deleting something that we don't
want to delete...

  "secrets is forbidden: User \"system:serviceaccount:pinniped:pinniped-config-controller-manager\" cannot deletecollection resource \"secrets\" in API group \"\" in the namespace \"default\"" "name"="w0" "namespace"="default" "reconciler group"="cluster.x-k8s.io" "reconciler kind"="Cluster"

Signed-off-by: Andrew Keesler <akeesler@vmware.com>